Add logVersionCard
